Closes #111.
Common types cannot trivially be used as this template is still using legacy version of metadata, but the main reason I had for logging the issue was that enums weren't getting their namespaces.

The problem was that the templates were "choosing" the wrong of two identically named extension methods, removing the using namespace for the wrong extension methods solved the problem.