Fix missing NuGet package causing `dotnet build` failures during SF run
This primarily fixes the problem that if a module is made with the Module Builder which does nothing except a `designer extension` with custom elements with extensions, then the `Intent.Modules.Common` NuGet package wasn't being installed which resulted in `dotnet build` failing during the SF run.

Otherwise it does minor neatening up in the affected files.